# MAINT: Block algorithm with a single copy per call to `block`