Block used to make repeated calls to concatenate. For a 3D block, this would copy the arrays 3 times. 1 time for each dimension of the block. This is 2 times too many.

This proposed algorithm does so with 1 memory copy.

It:

1. Calls `asanyarray` to make make the following logic easier.
2. Finds the shape of the final array
3. Finds the dtype of the final array.
4. Allocates the necessary memory.
5. Copies the memory in.

As is:
- [x] Passes all tests.
- [x] Code is clean
- [x] Code is commented
- [x] Improves on benchmark


For large arrays, the benchmark https://github.com/numpy/numpy/pull/11965 shows that there is now little room to improve compared to a straight concatenate operation.

The benchmarks show results for a final array of size `5 n`^3. Therefore, for n=10, the final array has 125,000 elements.

New results
```
===== ========== =============
--              mode          
----- ------------------------
n     block     concatenate 
===== ========== =============
1    130±5μs    2.57±0.07μs 
10   245±30μs     53.9±3μs  
100   503±4ms      555±9ms   
===== ========== =============
```

Improvement compared to the old results

```
asv continuous -E conda:3.6 -b Block.time_3d master block_single_concatenate_call 
```
```
       before           after         ratio
     [58da5687]       [66baab20]
     <master>         <block_single_concatenate_call>
+        49.4±1μs          130±5μs     2.64  bench_shape_base.Block.time_3d(1, 'block')
-      1.48±0.04s          503±4ms     0.34  bench_shape_base.Block.time_3d(100, 'block')
-      1.08±0.2ms         245±30μs     0.23  bench_shape_base.Block.time_3d(10, 'block')
```

Immediate things to improve on might be avoiding 3 calls to recursive functions. `asanyarray`, `shape` and `dtype` recursions can likely be combined.

Finally, this probably lends itself well to offering options like `order`, which would allow one to choose the order of the final array, or even to specify the array that they want to provide through an `out` parameter.


Closures are avoided entirely, so I  think that gh-10620 is not an issue.