@eric-wieser 2018-09-17
I had this in mind a while ago, and was hoping someone else would give it a go:)

This algorithm opens up the option of an "unblock" operation, which does exactly the same thing, but with the assignment in the other direction. Something to explore later.