@hmaarrfk 2018-09-17
> I'm afraid I introduced a conflict by merging #11910. The upshot is you could use closures again, if they help

no worries.