@hmaarrfk 2018-09-17
Suggests kept since they are being hidden by github:

@eric-wieser  wrt lines 447
Instead of doing the assignment, can you assemble a list of (slices, array) to return to the caller? That would let us implement unblock easily too. So this function would become _match_slices or similar

@eric-wieser 
Helper function for lines 457 to 465: result_shape = _concatenate_shape((shape1, shape2, ...), axis=axis), where all those shapes are just tuples

These require thinking and it is too late now for me to do that. Goodnight!

Thank you for your help.