@eric-wieser 2018-09-17
The latter shouldn't really require any work at all, just a copy-paste of the function contents. No rush though - thanks for the good work so far!

---

Regarding `unblock` - one of my original goals was to add python-style unpacking. Where in python we could write;

```
a, *b = range(5)
```

I was thinking of aiming for
```
a = np.empty(2)
b = np.empty(3)
np.b_[a, b] = np.arange(5)
```