@hmaarrfk 2018-09-17
I think you should make this unblock a separate issue.

We could do what you suggested, or what might also be useful is provided a list of shapes, We could then return views into the original array.

Do I need to comment the internal functions more than what I have now?

Personally, having gone through the super commented code of `pad`, I actually felt that the comments to internal functions were distracting. Especially since I couldn't see the signature and function body on the same screen...