@eric-wieser 2018-09-17
> having gone through the super commented code of pad, I actually felt that the comments to internal functions were distracting.

I agree

> Do I need to comment the internal functions more than what I have now?

Maybe a little more, but what you have looks pretty good. `unblock` should indeed be a separate issue - but if you can separate the slice computation from the assignment, then you can combine your `shape` recursion with the slice recursion.