@hmaarrfk 2018-09-17
I'm getting better performance thatn concatenate on large arrays..... is that a red herring?
```
[ 50.00%] ··· ===== ========= =============
              --              mode         
              ----- -----------------------
                n     block    concatenate 
              ===== ========= =============
                1    190±0μs     19.5±0μs  
                10   645±0μs     476±0μs   
               100   447±0ms     509±0ms   
              ===== ========= =============
```

htop says that assignment is single threaded :/ 

This is kinda consistent.

I think the poor performance on the 10 shaped array is due to the 2 extra unnecessary recursions (and probably the sanity checking code that could get cleaned up).