@hmaarrfk 2018-09-17
@eric-wieser That final refactoring is basically what is needed to get 

`_to_uniform_dim_recursion`
`_shape_recursion`
`_dtype_recursion`
`_assignment_recursion`

All in 1 recursion.

I was hoping to do it with some kind of list comprehension. I guess that won't happen though.