@eric-wieser 2018-09-17
To get it in a more comprehension-y form, I think you'd need to eliminate the `start_indices` argument entirely, and do repeated addition in the recursive step.