@hmaarrfk 2018-09-17
Honestly, list comprehension seems like a cop-out for python not to implement a JIT.

I understand generator expressions, but if they actually worked on a JIT, you wouldn't need to learn to read loops backward.

One day, I'll jump into PyPy.