@eric-wieser 2018-09-17
> Honestly, list comprehension seems like a cop-out for python not to implement a JIT.

I've really got no idea what you're talking about. As far as I'm aware, list comprehensions are nothing but (useful) syntactic sugar