@hmaarrfk 2018-09-17
Why was I convinced that list comprehensions were faster than loops. Eye opening!