@eric-wieser 2018-09-17
An OO-based approach to give you your comprehension-based recursive step:
```python
def _concatenate_offsets(shapes, axis):
    offset = 0
    offsets = []
    for shape in shape:
        offsets.append(offset)
        offset += shape[axis]
    return offsets


class concat_info:
    __slots__ = ('shape', 'dtype', 'locs')
    def __init__(self, shape, dtype, locs):
        self.shape = shape
        self.dtype = dtype
        self.locs = locs

    @classmethod
    def concatenate(cls, infos, axis):
        dtype = np.result_type(*(info.dtype for info in infos))
        shape = _concatenate_shapes([info.shape for info in infos], axis)
        offsets = _concatenate_offsets([info.shape for info in infos], axis)
        locs = [
            (
                loc[:axis] + (loc[axis] + offset,) + loc[axis+1:],
                arr
            )
            for info, offset in zip(infos, offsets)
            for loc, arr in info.locs
        ]
        return cls(shape, dtype)

    @classmethod
    def single(cls, arr):
        return concat_info(arr.shape, arr.dtype, [((0,)*arr.ndim, arr)])

    @property
    def regions(self):
        for loc, arr in self.locs:
            region = tuple(
                slice(start, start + size)
                for start, size in zip(loc, arr.shape)
            )
            yield region, arr

    def store(cls, out=None):
        if out is None:
            out = np.empty(self.shape, self.dtype)
        for index, arr in self.regions:
            out[index] = arr

    def load(cls, in_):
        if out is None:
            out = np.empty(self.shape, self.dtype)
        for index, arr in self.regions:
            arr[...] = in_[index]


def info_recurse(arrays, list_ndim, result_ndim, depth=0):
    if depth < list_ndim:
        axis = depth + result_ndim - list_ndim
        return concat_info.concatenate([
            info_recurse(arr, list_ndim, result_ndim, depth+1)
        ], axis=axis)
    else:
        return concat_info.single(arrays)
```

I don't know what the cost of the class is