@hmaarrfk 2018-09-17
Thanks Eric.

In my current state, I'm not really into OO these days. I always feel like it screams: "This is so complicated I first want you to read the documentation on this object".

I was able to slowly move things into a single recursion. Much easier now.

I've also implemented your regions so we can do unpacking really easily.
Somebody needs to write the docs and tests.