@hmaarrfk 2018-09-17
@eric-wieser I changed a test for the empty lists. This allows us to catch that fault early instead of having a special case in two places.