@hmaarrfk 2018-09-17
And if somebody wanted to implement `concatenate` on top of this new block function, they could do so with these lines of code inside `np.concatenate`

```python
    new_arrays = []
    for a in arrays:
        for _ in range(a.ndim - axis):
            a = [a]
        new_arrays.append(a)
    d = np.block(new_arrays)
```