@hmaarrfk 2018-09-17
Well I had fun creating a completely incomprehensible algorithm using list-comprehensions.