@eric-wieser 2018-09-17
Are you sure that switching to generators gives a performance boost? I'd worry it adds a lot of overhead here.