@hmaarrfk 2018-09-17
@eric-wieser it is really hard to tell the difference between the two implementestion (generator vs no generator). 

Honestly, I don't really care.