@hmaarrfk 2018-09-19
This commit: https://github.com/numpy/numpy/pull/11971/commits/9a70a86c59590223d5d26f175656801d47a5f233

Folded everything into a single recursion.

Unfortunately, it is still too slow. I guess list comprehensions and generators are just too slow compared to straight memory copy.

On my workstation, I still wasn't able to get good results bellow `512x512` arrays. For small arrays, this approach is still 2x slower. As a note, this didn't change much compared to the two recursion approach of commit https://github.com/numpy/numpy/pull/11971/commits/09ede13df130c152af271820ba20dca574ea5ba4

Though it might actually be slower! Details below.

My general sense is that small arrays are pretty important. The old approach was "wasteful" for large arrays, but quite quick for small arrays. There is probably room for both approaches.

By wisdom, I mean in the sense that FFTW uses "wisdom" to calculate the fastest method for computing a particular FFT on a particular machine (memory bound vs CPU bound).

<details>
<summary>Benchmark results compared to master </summary>

```
       before           after         ratio
     [9741ce23]       [7b0011ee]
     <master>         <block_single_concatenate_call>
+      21.2±0.2μs       40.1±0.2μs     1.90  bench_shape_base.Block2D.time_block2d((16, 16), 'uint64', (2, 2))
+      20.7±0.4μs         39.3±2μs     1.89  bench_shape_base.Block2D.time_block2d((16, 16), 'uint32', (2, 2))
+      20.7±0.3μs       39.1±0.2μs     1.89  bench_shape_base.Block2D.time_block2d((32, 32), 'uint8', (2, 2))
+      20.3±0.3μs         38.2±1μs     1.88  bench_shape_base.Block2D.time_block2d((16, 16), 'uint16', (2, 2))
+     21.9±0.09μs       41.1±0.9μs     1.88  bench_shape_base.Block2D.time_block2d((32, 32), 'uint64', (2, 2))
+      51.4±0.9μs       94.8±0.2μs     1.85  bench_shape_base.Block2D.time_block2d((16, 16), 'uint32', (4, 4))
+      51.4±0.6μs       94.4±0.3μs     1.84  bench_shape_base.Block2D.time_block2d((32, 32), 'uint8', (4, 4))
+        52.2±1μs         94.9±2μs     1.82  bench_shape_base.Block2D.time_block2d((16, 16), 'uint64', (4, 4))
+      52.9±0.5μs       94.9±0.2μs     1.79  bench_shape_base.Block2D.time_block2d((32, 32), 'uint32', (4, 4))
+      22.4±0.9μs       40.1±0.6μs     1.79  bench_shape_base.Block2D.time_block2d((32, 32), 'uint16', (2, 2))
+      22.1±0.4μs       39.5±0.5μs     1.79  bench_shape_base.Block2D.time_block2d((16, 16), 'uint8', (2, 2))
+        24.3±1μs       43.0±0.9μs     1.76  bench_shape_base.Block2D.time_block2d((64, 64), 'uint32', (2, 2))
+        23.3±1μs       41.1±0.6μs     1.76  bench_shape_base.Block2D.time_block2d((64, 64), 'uint16', (2, 2))
+        53.3±2μs         91.7±2μs     1.72  bench_shape_base.Block2D.time_block2d((16, 16), 'uint16', (4, 4))
+      54.7±0.7μs         93.5±2μs     1.71  bench_shape_base.Block2D.time_block2d((64, 64), 'uint16', (4, 4))
+      54.4±0.8μs         92.9±3μs     1.71  bench_shape_base.Block2D.time_block2d((32, 32), 'uint64', (4, 4))
+      54.9±0.5μs       93.6±0.7μs     1.71  bench_shape_base.Block2D.time_block2d((16, 16), 'uint8', (4, 4))
+      56.0±0.3μs         95.2±4μs     1.70  bench_shape_base.Block2D.time_block2d((32, 32), 'uint16', (4, 4))
+      23.1±0.2μs         39.0±1μs     1.69  bench_shape_base.Block2D.time_block2d((32, 32), 'uint32', (2, 2))
+      24.2±0.7μs       40.3±0.9μs     1.67  bench_shape_base.Block2D.time_block2d((64, 64), 'uint8', (2, 2))
+      59.6±0.6μs       97.4±0.2μs     1.63  bench_shape_base.Block2D.time_block2d((64, 64), 'uint32', (4, 4))
+        58.9±3μs       96.0±0.3μs     1.63  bench_shape_base.Block2D.time_block2d((128, 128), 'uint8', (4, 4))
+      63.3±0.8μs          101±3μs     1.59  bench_shape_base.Block2D.time_block2d((64, 64), 'uint64', (4, 4))
+      28.9±0.2μs       45.4±0.2μs     1.57  bench_shape_base.Block2D.time_block2d((128, 128), 'uint16', (2, 2))
+        70.6±2μs          108±5μs     1.53  bench_shape_base.Block2D.time_block2d((256, 256), 'uint8', (4, 4))
+      35.3±0.2μs       48.6±0.3μs     1.38  bench_shape_base.Block2D.time_block2d((128, 128), 'uint32', (2, 2))
+        83.6±4μs          115±4μs     1.38  bench_shape_base.Block2D.time_block2d((256, 256), 'uint16', (4, 4))
+        36.7±1μs       50.1±0.4μs     1.37  bench_shape_base.Block2D.time_block2d((256, 256), 'uint8', (2, 2))
+      83.3±0.6μs        111±0.4μs     1.33  bench_shape_base.Block2D.time_block2d((128, 128), 'uint64', (4, 4))
+        46.9±3μs       57.2±0.2μs     1.22  bench_shape_base.Block2D.time_block2d((256, 256), 'uint16', (2, 2))
+      45.8±0.2μs      54.3±0.09μs     1.18  bench_shape_base.Block2D.time_block2d((128, 128), 'uint64', (2, 2))
-       281±0.9μs          255±2μs     0.91  bench_shape_base.Block2D.time_block2d((1024, 1024), 'uint8', (4, 4))
-         238±1μs        205±0.5μs     0.86  bench_shape_base.Block2D.time_block2d((512, 512), 'uint32', (4, 4))
-         119±4μs         94.6±2μs     0.79  bench_shape_base.Block2D.time_block2d((512, 512), 'uint16', (2, 2))
-        424±30μs          322±3μs     0.76  bench_shape_base.Block2D.time_block2d((1024, 1024), 'uint16', (4, 4))
-       395±0.8μs          299±6μs     0.76  bench_shape_base.Block2D.time_block2d((512, 512), 'uint64', (4, 4))
-        209±60μs          149±4μs     0.71  bench_shape_base.Block2D.time_block2d((1024, 1024), 'uint8', (2, 2))
-        841±20μs         483±20μs     0.57  bench_shape_base.Block2D.time_block2d((1024, 1024), 'uint32', (4, 4))
-        6.56±2ms       1.18±0.1ms     0.18  bench_shape_base.Block2D.time_block2d((1024, 1024), 'uint64', (4, 4))
-        6.84±2ms      1.12±0.06ms     0.16  bench_shape_base.Block2D.time_block2d((1024, 1024), 'uint64', (2, 2))
-      1.31±0.5ms          215±4μs     0.16  bench_shape_base.Block2D.time_block2d((1024, 1024), 'uint16', (2, 2))
-      1.37±0.5ms          196±7μs     0.14  bench_shape_base.Block2D.time_block2d((512, 512), 'uint64', (2, 2))
-        2.93±1ms         338±10μs     0.12  bench_shape_base.Block2D.time_block2d((1024, 1024), 'uint32', (2, 2))
```
<details>

<details>
<summary>Compared to 2 recusion approach</summary>

```
       before           after         ratio
     [09ede13d]       [7b0011ee]
     <block_single_concatenate_call~16>       <block_single_concatenate_call>
+        35.3±2μs       41.9±0.2μs     1.19  bench_shape_base.Block2D.time_block2d((128, 128), 'uint8', (2, 2))
+      33.8±0.5μs         39.5±1μs     1.17  bench_shape_base.Block2D.time_block2d((64, 64), 'uint8', (2, 2))
+        87.1±1μs        102±0.4μs     1.17  bench_shape_base.Block2D.time_block2d((128, 128), 'uint16', (4, 4))
+        38.7±1μs       44.6±0.5μs     1.15  bench_shape_base.Block2D.time_block2d((128, 128), 'uint16', (2, 2))
+        37.0±1μs      42.0±0.07μs     1.14  bench_shape_base.Block2D.time_block2d((64, 64), 'uint32', (2, 2))
+     36.5±0.09μs       41.4±0.3μs     1.13  bench_shape_base.Block2D.time_block2d((64, 64), 'uint16', (2, 2))
+        35.7±1μs         39.9±1μs     1.12  bench_shape_base.Block2D.time_block2d((32, 32), 'uint64', (2, 2))
+         114±6μs          126±5μs     1.11  bench_shape_base.Block2D.time_block2d((256, 256), 'uint32', (4, 4))
+        68.2±2μs       75.3±0.6μs     1.10  bench_shape_base.Block2D.time_block2d((512, 512), 'uint8', (2, 2))
```
<details>